[fix] usr_lib: 'better' code + more size_t instead of uint64_t

+++ usr_lib/kgui/widget.cpp
@@ -5,16 +5,16 @@ namespace gui {
 
     widget::widget(){
     }
-    void widget_list::init(uint64_t length){
+    void widget_list::init(size_t length){
         list = (gui::widget**)sys::service_malloc(length * sizeof (uint64_t));
-        for(uint64_t i = 0; i < length; i++){
+        for(size_t i = 0; i < length; i++){
             list[i] = nullptr;
         }
         list_length = length;
     }
 
     void widget_list::add_widget(widget* widget){
-        for(uint64_t i = 0; i < list_length; i++){
+        for(size_t i = 0; i < list_length; i++){
             if(list[i] == nullptr){
                 list[i] = widget;
                 return;
@@ -24,7 +24,7 @@ namespace gui {
 
     void widget_list::update_all(){
 
-        for(uint64_t i = 0; i < list_length; i++){
+        for(size_t i = 0; i < list_length; i++){
             if(list[i] != nullptr){
                 list[i]->update_widget();
             }
@@ -32,7 +32,7 @@ namespace gui {
     }
     void widget_list::draw_all(sys::graphic_context& context){
 
-        for(uint64_t i = 0; i < list_length; i++){
+        for(size_t i = 0; i < list_length; i++){
             if(list[i] != nullptr){
                 list[i]->draw_widget(context);
             }
